requirements.txt
# Core ML libraries
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Statistics
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# CLI
click>=8.1.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing (optional)
pytest>=7.4.0
